requirements.txt
customtkinter
screeninfo
pyautogui
pyserial